use mocha/chai for int tests
