+++ integration-tests/server/.babelrc
@@ -1,5 +1,4 @@
 {
-  "presets": [
-    "@babel/env"
-  ]
+  "plugins": ["@babel/plugin-proposal-object-rest-spread"],
+  "presets": ["@babel/env"]
 }
