+++ integration-tests/server/package.json
@@ -2,7 +2,7 @@
   "name": "integration-tests-server",
   "private": "true",
   "scripts": {
-    "test": "jest"
+    "test": "mocha --require @babel/register"
   },
   "dependencies": {
     "@arranger/server": "^0.1.35",
@@ -12,10 +12,14 @@
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-beta.40",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0-beta.40",
     "@babel/preset-env": "^7.0.0-beta.40",
+    "@babel/register": "^7.0.0-beta.40",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^22.2.2",
+    "chai": "^4.1.2",
     "jest": "^22.4.2",
+    "mocha": "^5.0.4",
     "regenerator-runtime": "^0.11.1"
   },
   "version": "0.1.35"
