+++ modules/server/src/projects/addProject.js
@@ -18,8 +18,6 @@ export default async (req, res) => {
     },
   };
 
-  console.log('wat', es.create);
-
   try {
     await es.create({
       ...arrangerconfig.projectsIndex,
